+++ Firebase/Storage/Private/FIRStorageGetDownloadURLTask.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Google
+ * Copyright 2018 Google
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,21 @@
  * limitations under the License.
  */
 
-#import "FIRCollectionReference.h"
-#import "FIRDocumentChange.h"
-#import "FIRDocumentReference.h"
-#import "FIRDocumentSnapshot.h"
-#import "FIRFieldPath.h"
-#import "FIRFieldValue.h"
-#import "FIRFirestore.h"
-#import "FIRFirestoreErrors.h"
-#import "FIRFirestoreSettings.h"
-#import "FIRGeoPoint.h"
-#import "FIRListenerRegistration.h"
-#import "FIRQuery.h"
-#import "FIRQuerySnapshot.h"
-#import "FIRSnapshotMetadata.h"
-#import "FIRTimestamp.h"
-#import "FIRTransaction.h"
-#import "FIRWriteBatch.h"
+#import "FIRStorageTask.h"
+
+@class GTMSessionFetcherService;
+
+NS_ASSUME_NONNULL_BEGIN
+
+/**
+ * Task which provides the ability to get a download URL for an object in Firebase Storage.
+ */
+@interface FIRStorageGetDownloadURLTask : FIRStorageTask <FIRStorageTaskManagement>
+
+- (instancetype)initWithReference:(FIRStorageReference *)reference
+                   fetcherService:(GTMSessionFetcherService *)service
+                       completion:(FIRStorageVoidURLError)completion;
+
+@end
+
+NS_ASSUME_NONNULL_END