+++ Firebase/Storage/Private/FIRStorageGetDownloadURLTask_Private.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Google
+ * Copyright 2018 Google
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,19 +14,17 @@
  * limitations under the License.
  */
 
-#import "FIRSetOptions.h"
+#import "FIRStorageGetDownloadURLTask.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface FIRSetOptions ()
+/**
+ * Task which provides the ability to get a download URL for an object in Firebase Storage.
+ */
+@interface FIRStorageGetDownloadURLTask ()
 
-- (instancetype)initWithMerge:(BOOL)merge NS_DESIGNATED_INITIALIZER;
-
-@end
-
-@interface FIRSetOptions (Internal)
-
-+ (instancetype)overwrite;
+/** Extracts a download URL from the StorageMetadata dictonary representation. */
++ (nullable NSURL *)downloadURLFromMetadataDictionary:(NSDictionary *)dictionary;
 
 @end